+++ app/containers/mainPanel.js
@@ -233,11 +233,9 @@ class MainPanel extends Component {
                 </div>
               </div>
               <div className='channel-meta__other'>
-                <div
-                  className='channel-meta__other__members'
-                  onClick={toggleMemberList}
-                >
-                  <img src='static/images/user-icon.svg' /> {channelMemberCount}
+                <div className='channel-meta__other__members' onClick={toggleMemberList}>
+                  <img src='static/images/user-icon.svg' />
+                  <div className='channel-meta__other__members__count'>{channelMemberCount}</div>
                 </div>
                 <div
                   onClick={this.showCabalSettings.bind(this, this.props.addr)}